Update Contentful types Don't pull navbar from ctf just for the logo.
